Refactor bitbucket-pipeline to use own image
First step of the build (after Dockerlint) now assembles a Docker image
containing all the airflow stuff. Every subsequent step (linting, etc.) uses
this image. Docker push only happens on master, and only if all checks are
successful.

We may want to change this so that the docker push with the build-number tag
happens for all branches, and only the docker-push for the `master` branch
includes the `:latest` tag.

I don't expect this pipeline to succeed on the first attempt. We'll see.